Broke BACnet application class into own file.
